refactor: 상품 Entity 를 외부에서 DTO 로 변경하여 주입하도록 변경
