+++ backend/src/main/java/com/team28/sidedish/service/ProductDiscountService.java
@@ -1,11 +1,13 @@
 package com.team28.sidedish.service;
 
 import com.team28.sidedish.controller.dto.ProductDetailResponse;
+import com.team28.sidedish.controller.dto.ProductImageResponse;
 import com.team28.sidedish.controller.dto.ProductListResponse;
 import com.team28.sidedish.controller.dto.ProductResponse;
 import com.team28.sidedish.domain.Product;
 import com.team28.sidedish.repository.entity.DiscountEntity;
 import com.team28.sidedish.repository.entity.ProductEntity;
+import com.team28.sidedish.repository.entity.ProductImageEntity;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -45,6 +47,16 @@ public class ProductDiscountService {
 
         Product product = Product.of(productEntity, discountEntity);
 
-        return ProductDetailResponse.from(product);
+        return ProductDetailResponse.of(
+                product,
+                convertToProductImageResponses(product.getRepresentImages()),
+                convertToProductImageResponses(product.getDetailImages())
+        );
+    }
+
+    private static List<ProductImageResponse> convertToProductImageResponses(List<ProductImageEntity> images) {
+        return images.stream()
+                .map(ProductImageResponse::from)
+                .collect(Collectors.toUnmodifiableList());
     }
 }
